requirement.txt
Django==3.1
psycopg2-binary==2.9.7
boto3
numpy
django-allauth
django_extensions
django-celery
django-cors-headers
django-debug-toolbar
djangorestframework
django-appconf
django-cachalot
django-compressor
django-crequest
django-dbbackup
django-extra-views
django-jinja
django-jsonfield
django-modeltranslation
django-redis-cache
django-reversion
django-silk
django-storages
django-tables2
django-taggit
django-widget-tweaks
celery
docutils
dropbox
et-xmlfile
FormEncode
gprof2dot
gunicorn
idna
imagesize
inflect
isort
isoweek
jdcal
Jinja2
jmespath
kombu
lazy-object-proxy
linecache2
MarkupSafe
mccabe
mongoengine
newrelic
oauthlib
olefile
openpyxl
packaging
pathlib
pycodestyle
Pygments
pylint
pymongo
pyparsing
python-dateutil
python-memcached
python3-openid
pytz
pyxform
PyYAML
rcssmin
redis
requests
requests-oauthlib
rjsmin
s3transfer
savReaderWriter==3.4.2
setproctitle
six
snowballstemmer
sqlparse
typing
unicodecsv
Unidecode
unittest2
urllib3
vine
wrapt
xlrd
xlwt